+++ main.py
@@ -1,0 +1,5 @@
+from resources.password import generate_random_password
+
+## invoke the function
+password = generate_random_password(6,1,1)
+print('Password:', password)